+++ src/AddPerceptronTest.java
@@ -61,34 +61,34 @@ public class AddPerceptronTest {
 		Perceptron p2 = ap.getP2();
 		Perceptron p3 = ap.getP3();
 		
-		assertEquals(p1.getIn1(), 1);
-		assertEquals(p1.getIn2(), 0);
+		assertEquals((int)p1.getIn1(), 1);
+		assertEquals((int)p1.getIn2(), 0);
 		
-		assertEquals(p2.getIn1(), 1);
-		assertEquals(p2.getIn2(), 1);
+		assertEquals((int)p2.getIn1(), 1);
+		assertEquals((int)p2.getIn2(), 1);
 		
-		assertEquals(p3.getIn1(), 1);
-		assertEquals(p3.getIn2(), 0);
+		assertEquals((int)p3.getIn1(), 1);
+		assertEquals((int)p3.getIn2(), 0);
 		
 	}
 	
 	@Test
 	public void addTest(){
 		ap.feed(0, 0);
-		assertEquals(ap.getSum(), 0);
-		assertEquals(ap.getCarry(), 0);
+		assertEquals((int)ap.getSum(), 0);
+		assertEquals((int)ap.getCarry(), 0);
 		
 		ap.feed(0, 1);
-		assertEquals(ap.getSum(), 1);
-		assertEquals(ap.getCarry(), 0);
+		assertEquals((int)ap.getSum(), 1);
+		assertEquals((int)ap.getCarry(), 0);
 		
 		ap.feed(1, 0);
-		assertEquals(ap.getSum(), 1);
-		assertEquals(ap.getCarry(), 0);
+		assertEquals((int)ap.getSum(), 1);
+		assertEquals((int)ap.getCarry(), 0);
 		
 		ap.feed(1, 1);
-		assertEquals(ap.getSum(), 0);
-		assertEquals(ap.getCarry(), 1);
+		assertEquals((int)ap.getSum(), 0);
+		assertEquals((int)ap.getCarry(), 1);
 
 	}
